move container files into kiosk module
