+++ src/Kiosk/Kiosk.jsx
@@ -1,17 +1,21 @@
 import React from 'react'
 import VisibleItemList from './VisibleItemList'
+import CategoryContainer from './CategoryContainer'
+import VisibleCartList from './VisibleCartList'
 
-const Kitchen = React.createClass({
+const Kiosk = React.createClass({
     render: function () {
         return (
             <div>
                 <div id='content' className='row pos-container'>
+                    <CategoryContainer />
                     <VisibleItemList />
+                    <VisibleCartList />
                 </div>
             </div>
         )
     }
 })
 
-export default Kitchen
+export default Kiosk
